requirements.txt
streamlit
google-generativeai
python-dotenv
google-api-python-client
PyPDF2
Pillow